Changed amount field from decimal to float

+++ arms_sales/models.py
@@ -5,7 +5,7 @@ class Proposed(models.Model):
     title = models.CharField(max_length=200)
     text = models.TextField(null=True, blank=True)
     date = models.DateField(null=True, blank=True)
-    amount = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
+    amount = models.FloatField(null=True, blank=True)
     location = models.CharField(max_length=200)
     location_id = models.IntegerField(null=True, blank=True)
     # from DSCAmodels.IntegerField(
